API/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
fonttools==4.49.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-version==0.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
